pyproject.toml
[tool.poetry]
name = "graphrag-kit"
version = "0.1.0"
description = ""
authors = ["Theodore Niu <theodore.niu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = ">=1.0.1,<1.1.0"
pyodbc = "5.1.0"
streamlit = "1.39.0"
streamlit-chat = "0.1.1"
qrcode = "7.4.2"
psycopg2-binary = "2.9.9"
future = "1.0.0"
streamlit-javascript = "0.1.5"
langchain-community = "0.3.3"
openpyxl = "3.1.5"
pymupdf = "1.24.12"
pdfplumber = "0.11.4"
pypdf2 = "3.0.1"
xlrd = "2.0.1"
fastapi = {version = "0.115.4", extras = ["standard"]}
uvicorn = "0.32.0"
pydantic = ">=2.9.2,<2.10.0"
requests = ">=2.32.3,<2.33.0"
openai = "1.53.0"
httpx = "0.27.2"
streamlit-authenticator = "0.4.1"
azure-ai-documentintelligence = "1.0.0b4"
theodoretools = "0.0.8"
azure-ai-formrecognizer = "3.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
